+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add support for MySQL 8.4
 - ignore extraneous vscode triggered updates in watched directory
 - add support for mp4 responses
+- add support for zoneinfo in Python 3.9+
 
 ## [6.23.0] - 2024-04-24
 - add site timezone attribute